+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 - Shift Type Debouncer customizable using new NumericType trait
 - use strategy Active specification from associated function instead of struct
+- add new Debounced creation options
 
 ## [0.1.1] - 2020-11-13
 ### Added