+++ CHANGELOG.md
@@ -1,20 +1,23 @@
 # Changelog
 
 ## [Unreleased]
+
+## [0.2.0]—2020-12-27
 - Shift Type Debouncer customizable using new NumericType trait
 - use strategy Active specification from associated function instead of struct
 - add new Debounced creation options
 
-## [0.1.1] - 2020-11-13
+## [0.1.1]—2020-11-13
 ### Added
 - This CHANGELOG
 - Complete documentation (still lacking examples)
 - Add function to compute number of samples needed for a hold time
 
-## [0.1.0] - 2020-11-01
+## [0.1.0]—2020-11-01
 ### Added
 - Initial implementation of integrator strategy and type hierarchy
 
-[Unreleased]: https://github.com/brendanarciszewski/debounced-signals/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/brendanarciszewski/debounced-signals/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/brendanarciszewski/debounced-signals/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/brendanarciszewski/debounced-signals/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/brendanarciszewski/debounced-signals/releases/tag/v0.1.0